# docs/swagger.yaml
definitions:
...
  DatabaseServicev1.CreateWardRequest:
    properties:
      address:
        description: '* Адрес подопечного'
        type: string
      fullName:
        description: '* Полное имя подопечного'
        type: string
      necessary:
        description: '* Необходимая сумма денег на необходимость'
        type: number
      title:
        description: '* Дополнительный текст к подопечному'
        type: string
      want:
        description: '* Необходимость подопечного (то в чем он нуждается, например
          "Лекарства")'
        type: string
    type: object
...
  DatabaseServicev1.Donations:
    properties:
      amount:
        description: '* Сумма пожертвования'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID пожертвования в базе данных'
        type: integer
      title:
        description: '* Название пожертвования (например "На лекарства")'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя, которому принадлежит пожертвование'
        type: integer
      wardId:
        description: '* ID подопечного для которого предназначено данное пожертвование'
        type: integer
    type: object
  DatabaseServicev1.DonationsResponse:
    properties:
      donations:
        description: '* Массив пожертвований'
        items:
          $ref: '#/definitions/DatabaseServicev1.Donations'
        type: array
    type: object
...
  DatabaseServicev1.HTTPCodes:
    properties:
      code:
        description: '* HTTP код ответа'
        type: integer
    type: object
...
  DatabaseServicev1.Ward:
    properties:
      address:
        description: '* Адрес подопечного'
        type: string
      collected:
        description: '* Количество собранных средств'
        type: number
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      donations:
        description: '* пожертвования для подопечного'
        items:
          $ref: '#/definitions/DatabaseServicev1.Donations'
        type: array
      fullName:
        description: '* Полное имя подопечного'
        type: string
      id:
        description: '* ID подопечного в базе данных'
        type: integer
      necessary:
        description: '* Необходимая сумма денег на потребность'
        type: number
      title:
        description: '* Дополнительный текст к подопечному'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      want:
        description: '* Потребность подопечного (то в чем он нуждается, например "Лекарства")'
        type: string
    type: object
  DatabaseServicev1.WardsResponse:
    properties:
      wards:
        description: '* Массив подопечных'
        items:
          $ref: '#/definitions/DatabaseServicev1.Ward'
        type: array
    type: object
  server.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Сервер маршрутизации
  title: API Gateway
  version: "1.0"
paths:
...
  /api/v1/wards:
    get:
      consumes:
      - application/json
      description: Список всех подопечных в базе данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.WardsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Список всех подопечных
      tags:
      - Wards
    post:
      consumes:
      - application/json
      description: Создание подопечного
      parameters:
      - description: Сущность подопечного
        in: body
        name: ward
        schema:
          $ref: '#/definitions/DatabaseServicev1.CreateWardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Ward'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Создание подопечного
      tags:
      - Wards
    put:
      consumes:
      - application/json
      description: Обновление подопечного
      parameters:
      - description: Модель для обновления
        in: body
        name: ward
        schema:
          $ref: '#/definitions/DatabaseServicev1.Ward'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Ward'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновление подопечного
      tags:
      - Wards
  /api/v1/wards/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление подопечного по ID
      parameters:
      - description: ID подопечного
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление подопечного
      tags:
      - Wards
    get:
      consumes:
      - application/json
      description: Поиск подопечного по ID
      parameters:
      - description: ID подопечного
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Ward'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Поиск подопечного
      tags:
      - Wards
  /api/v1/wards/{id}/donations:
    get:
      consumes:
      - application/json
      description: Извлечение пожертвований подопечного по его ID
      parameters:
      - description: Ward ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.DonationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Извлечение пожертвований подопечного
      tags:
      - Wards
  /api/v1/wards/deleteModel:
    post:
      consumes:
      - application/json
      description: Удаляет подопечного опираясь на всю сущность модели
      parameters:
      - description: Сущность подопечного
        in: body
        name: ward
        schema:
          $ref: '#/definitions/DatabaseServicev1.Ward'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление подопечного по модели
      tags:
      - Wards
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"